Header, player and Botton banner
In this version I edit some bugs on the header (the player was fixed
and now is absolute when window > 1000).

Also I add a bottom banner. With some localStorage capabilites (in
order to show or not show the banner to the user).

Finally I change the color of the template